backend/automation/requirements.txt
asgiref==3.11.1
Django==5.2.11
djangorestframework==3.16.1
mssql-django==1.6
pyodbc==5.3.0
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3
